api/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
supabase==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
moviepy==1.0.3
pydantic==2.4.2
pydantic-settings==2.0.3
google-generativeai
scenedetect
opencv-python